Update architecture
- Created tests folder to separate test modules per file
